Do not distribute debian-standard schema files
(This is possible because we keep everything Viper-related in spinlock.schema,
and for a couple existing objectclasses we need to change, we introduce x- (private)
version of them and and use that in LDIF files).

Later I'll see what's the right approach around all this (the set of supported
attrs, OCs etc.).